feat: refactor chart type utilities and constants for better organization
