+++ packages/mosaic/src/chart-builders/chartTypeUtils.ts
@@ -3,39 +3,14 @@ import type {
   ChartBuilderField,
   ChartTypeDefinition,
 } from './types';
+import {buildDefaultChartTitle as buildDefaultChartTitleImpl} from './constants';
 
-export const NUMERIC_COLUMN_TYPES = [
-  'BIGINT',
-  'BIT',
-  'DECIMAL',
-  'DOUBLE',
-  'FLOAT',
-  'HUGEINT',
-  'INTEGER',
-  'REAL',
-  'SMALLINT',
-  'TINYINT',
-  'UBIGINT',
-  'UHUGEINT',
-  'UINTEGER',
-  'USMALLINT',
-  'UTINYINT',
-];
-
-export const TEMPORAL_COLUMN_TYPES = [
-  'DATE',
-  'TIME',
-  'TIMESTAMP',
-  'TIMESTAMP_MS',
-  'TIMESTAMP_NS',
-  'TIMESTAMP_S',
-  'TIMESTAMPTZ',
-];
-
-export const QUANTITATIVE_COLUMN_TYPES = [
-  ...NUMERIC_COLUMN_TYPES,
-  ...TEMPORAL_COLUMN_TYPES,
-];
+export {
+  NUMERIC_COLUMN_TYPES,
+  TEMPORAL_COLUMN_TYPES,
+  QUANTITATIVE_COLUMN_TYPES,
+  buildDefaultChartTitle,
+} from './constants';
 
 export function columnMatchesFieldTypes(
   column: ChartBuilderColumn,
@@ -76,25 +51,13 @@ export function getAvailableChartTypes(
   );
 }
 
-export function buildDefaultChartTitle(
-  description: string,
-  fieldValues: Record<string, string>,
-): string {
-  const baseTitle = description.replace(/^Create (a |an )?/, '');
-  const selectedFields = Object.values(fieldValues).filter(Boolean);
-
-  return selectedFields.length > 0
-    ? `${baseTitle} - ${selectedFields.join(', ')}`
-    : baseTitle;
-}
-
 export function buildChartTypeTitle(
   chartType: Pick<ChartTypeDefinition, 'description' | 'buildTitle'>,
   fieldValues: Record<string, string>,
 ): string {
   return chartType.buildTitle
     ? chartType.buildTitle(fieldValues)
-    : buildDefaultChartTitle(chartType.description, fieldValues);
+    : buildDefaultChartTitleImpl(chartType.description, fieldValues);
 }
 
 export function canCreateChartFromType(
