+++ packages/mosaic/src/chart-types/base-types.ts
@@ -1,0 +1,17 @@
+/**
+ * Base types for chart configurations.
+ * Separated to avoid circular dependencies.
+ */
+
+/**
+ * Supported chart type identifiers
+ */
+export type VgPlotChartType =
+  | 'histogram'
+  | 'count-plot'
+  | 'ecdf'
+  | 'line-chart'
+  | 'bubble-chart'
+  | 'heatmap'
+  | 'box-plot'
+  | 'custom-spec';
