+++ packages/mosaic/src/chart-types/custom-spec/definition.ts
@@ -1,7 +1,7 @@
 import type {Spec} from '@uwdata/mosaic-spec';
 import type {ChartTypeDefinition} from '../../chart-builders/types';
 import type {CustomSpecChartSettings} from './schema';
-import {buildDefaultChartTitle} from '../../chart-builders/chartTypeUtils';
+import {buildDefaultChartTitle} from '../../chart-builders/constants';
 
 function titleFromDescription(description: string) {
   return (fieldValues: Record<string, string>) =>
