+++ packages/mosaic/src/chart-types/ecdf/definition.ts
@@ -4,7 +4,7 @@ import type {EcdfChartSettings} from './schema';
 import {
   buildDefaultChartTitle,
   QUANTITATIVE_COLUMN_TYPES,
-} from '../../chart-builders/chartTypeUtils';
+} from '../../chart-builders/constants';
 
 const FG_COLOR = 'var(--color-chart-1)';
