+++ packages/mosaic/src/chart-types/line-chart/definition.ts
@@ -5,7 +5,7 @@ import {
   buildDefaultChartTitle,
   QUANTITATIVE_COLUMN_TYPES,
   NUMERIC_COLUMN_TYPES,
-} from '../../chart-builders/chartTypeUtils';
+} from '../../chart-builders/constants';
 
 const FG_COLOR = 'var(--color-chart-1)';
